add roles and permissions

+++ app/Http/Controllers/Admin/ExpenseCategoryController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\{StoreExpenseCategoryFormRequest, UpdateExpenseCategoryFormRequest};
 use App\Repositories\ExpenseCategoryRepository;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class ExpenseCategoryController extends Controller
 {
@@ -15,12 +16,16 @@ class ExpenseCategoryController extends Controller
     }
     public function index()
     {
+        abort_if(Gate::denies('view_category'), Response::HTTP_FORBIDDEN);
+        
         $expenseCategorys = $this->expenseCategory->getAll();
 
         return $this->successResponse($expenseCategorys);
     }
     public function store(StoreExpenseCategoryFormRequest $request)
     {
+        abort_if(Gate::denies('create_category'), Response::HTTP_FORBIDDEN);
+
         $validated = $request->validated();
 
         $expenseCategory = $this->expenseCategory->create($validated);
@@ -29,10 +34,14 @@ class ExpenseCategoryController extends Controller
     }
     public function show($expenseCategory)
     {   
+        abort_if(Gate::denies('view_category'), Response::HTTP_FORBIDDEN);
+
         return $this->successResponse($this->expenseCategory->getById($expenseCategory));
     }
     public function update(UpdateExpenseCategoryFormRequest $request, $expenseCategory)
     {   
+        abort_if(Gate::denies('update_category'), Response::HTTP_FORBIDDEN);
+
         $validated = $request->validated();
        
         $expenseCategory = $this->expenseCategory->update($validated, $expenseCategory);
@@ -41,6 +50,8 @@ class ExpenseCategoryController extends Controller
     }
     public function destory($expenseCategory) 
     {
+        abort_if(Gate::denies('delete_category'), Response::HTTP_FORBIDDEN);
+
         $expenseCategory = $this->expenseCategory->delete($expenseCategory);
 
         return $this->successResponse($expenseCategory);
